Negated BATTSCI current value
Also changed how BATTSCI sends frames (it now ping-pongs between the two).  Added new CAR about how MCM is suddenly drooping Connector E voltage... I guess it behaves differently when actually installed in the car (and all the signals are present... so no P-codes).  We can certainly fix this issue in firmware (e.g. by spoofing voltage down to match)... but a better solution is to add more capacitance and/or lower the resistance to the spoofing voltage divider.