Update battsci.cpp
a large change in SOC from gridcharging meant that the battsci value jumped and missed out key setpoints. (eg grid charge from 15% to 82% and background regen was still enabled)

with this change previousOutputSoC_deciPercent is set to 65%SOC at key-on, therefore all appropriate setpoints are triggered shortly after keyon. In the above example, SOC 72% would be triggered and therefore disable bg regen.

If the pack had decreased since last key on, bg regen would be triggered if below 60%SOC or 15% if bg regen is disabled.